REFACTOR: one main cpp file for all challenges
now pass the day and file on CLI to simplify codebase
